# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Meta Ads Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /adsets:
    get:
      operationId: listAdSets
      tags: [adsets]
      summary: List ad sets
      parameters:
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Ad set list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdSet"
    post:
      operationId: createAdSet
      tags: [adsets]
      summary: Create an ad set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdSetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdSet"
  /adsets/{id}:
    get:
      operationId: getAdSet
      tags: [adsets]
      summary: Get an ad set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdSet"
    patch:
      operationId: updateAdSet
      tags: [adsets]
      summary: Update an ad set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdSetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdSet"
    delete:
      operationId: deleteAdSet
      tags: [adsets]
      summary: Delete an ad set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /adsets/{id}/toggle:
    post:
      operationId: toggleAdSet
      tags: [adsets]
      summary: Toggle ad set on/off
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdSet"
  /adsets/{id}/duplicate:
    post:
      operationId: duplicateAdSet
      tags: [adsets]
      summary: Duplicate an ad set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdSet"

  # ─── ADS ─────────────────────────────────────────────────────────
  /ads:
    get:
      operationId: listAds
      tags: [ads]
      summary: List ads
      parameters:
        - name: adsetId
          in: query
          schema:
            type: ["integer", "null"]
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Ads list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      operationId: createAd
      tags: [ads]
      summary: Create an ad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
  /ads/{id}:
    get:
      operationId: getAd
      tags: [ads]
      summary: Get an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
    patch:
      operationId: updateAd
      tags: [ads]
      summary: Update an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
    delete:
      operationId: deleteAd
      tags: [ads]
      summary: Delete an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /ads/{id}/toggle:
    post:
      operationId: toggleAd
      tags: [ads]
      summary: Toggle ad on/off
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
  /ads/{id}/duplicate:
    post:
      operationId: duplicateAd
      tags: [ads]
      summary: Duplicate an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"

  # ─── INSIGHTS / SUMMARY ──────────────────────────────────────────
...
components:
  schemas:
...
    AdSet:
      type: object
      required: [id, campaignId, campaignName, name, status, delivery, amountSpent, impressions, reach, clicks, results, costPerResult, cpm, ctr, frequency, startDate, createdAt]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        campaignName:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [ACTIVE, PAUSED, DELETED, ARCHIVED]
        delivery:
          type: string
          enum: [active, learning, off, error, not_delivering]
        budgetType:
          type: string
          enum: [daily, lifetime]
        budget:
          type: ["number", "null"]
        amountSpent:
          type: number
        impressions:
          type: integer
        reach:
          type: integer
        clicks:
          type: integer
        results:
          type: integer
        costPerResult:
          type: ["number", "null"]
        cpm:
          type: ["number", "null"]
        ctr:
          type: ["number", "null"]
        frequency:
          type: ["number", "null"]
        audience:
          type: ["string", "null"]
        placement:
          type: ["string", "null"]
        optimization:
          type: ["string", "null"]
        billingEvent:
          type: ["string", "null"]
        conversionLocation:
          type: ["string", "null"]
        facebookPage:
          type: ["string", "null"]
        performanceGoal:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        ageMin:
          type: ["integer", "null"]
        ageMax:
          type: ["integer", "null"]
        languages:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        detailedTargeting:
          type: ["string", "null"]
        placementMode:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        recommendation:
          type: ["string", "null"]
        createdAt:
          type: string
    AdSetInput:
      type: object
      required: [campaignId, name, budgetType]
      properties:
        campaignId:
          type: integer
        name:
          type: string
        budgetType:
          type: string
          enum: [daily, lifetime]
        budget:
          type: ["number", "null"]
        amountSpent:
          type: ["number", "null"]
        costPerResult:
          type: ["number", "null"]
        audience:
          type: ["string", "null"]
        placement:
          type: ["string", "null"]
        optimization:
          type: ["string", "null"]
        conversionLocation:
          type: ["string", "null"]
        facebookPage:
          type: ["string", "null"]
        performanceGoal:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        ageMin:
          type: ["integer", "null"]
        ageMax:
          type: ["integer", "null"]
        languages:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        detailedTargeting:
          type: ["string", "null"]
        placementMode:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
    AdSetUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        budgetType:
          type: ["string", "null"]
        budget:
          type: ["number", "null"]
        amountSpent:
          type: ["number", "null"]
        costPerResult:
          type: ["number", "null"]
        audience:
          type: ["string", "null"]
        conversionLocation:
          type: ["string", "null"]
        facebookPage:
          type: ["string", "null"]
        performanceGoal:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        ageMin:
          type: ["integer", "null"]
        ageMax:
          type: ["integer", "null"]
        languages:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        detailedTargeting:
          type: ["string", "null"]
        placementMode:
          type: ["string", "null"]
        placement:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]

    # ── Ad ────────────────────────────────────────────────────────
    Ad:
      type: object
      required: [id, adsetId, adsetName, campaignId, campaignName, name, status, delivery, format, amountSpent, impressions, reach, clicks, results, costPerResult, cpm, ctr, frequency, createdAt]
      properties:
        id:
          type: integer
        adsetId:
          type: integer
        adsetName:
          type: string
        campaignId:
          type: integer
        campaignName:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [ACTIVE, PAUSED, DELETED, ARCHIVED]
        delivery:
          type: string
          enum: [active, learning, off, error, not_delivering]
        format:
          type: string
          enum: [image, video, carousel, collection, stories]
        headline:
          type: ["string", "null"]
        primaryText:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        callToAction:
          type: ["string", "null"]
        destinationUrl:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        identityPage:
          type: ["string", "null"]
        instagramHandle:
          type: ["string", "null"]
        amountSpent:
          type: number
        impressions:
          type: integer
        reach:
          type: integer
        clicks:
          type: integer
        results:
          type: integer
        costPerResult:
          type: ["number", "null"]
        cpm:
          type: ["number", "null"]
        ctr:
          type: ["number", "null"]
        frequency:
          type: ["number", "null"]
        recommendation:
          type: ["string", "null"]
        createdAt:
          type: string
    AdInput:
      type: object
      required: [adsetId, name, format]
      properties:
        adsetId:
          type: integer
        name:
          type: string
        format:
          type: string
          enum: [image, video, carousel, collection, stories]
        amountSpent:
          type: ["number", "null"]
        costPerResult:
          type: ["number", "null"]
        headline:
          type: ["string", "null"]
        primaryText:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        callToAction:
          type: ["string", "null"]
        destinationUrl:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        identityPage:
          type: ["string", "null"]
        instagramHandle:
          type: ["string", "null"]
    AdUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        format:
          type: ["string", "null"]
        amountSpent:
          type: ["number", "null"]
        costPerResult:
          type: ["number", "null"]
        headline:
          type: ["string", "null"]
        primaryText:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        callToAction:
          type: ["string", "null"]
        destinationUrl:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        identityPage:
          type: ["string", "null"]
        instagramHandle:
          type: ["string", "null"]

    # ── Insights ──────────────────────────────────────────────────
...