# lib/api-zod/src/generated/types/adInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Meta Ads Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AdInputFormat } from './adInputFormat';

export interface AdInput {
  adsetId: number;
  name: string;
  format: AdInputFormat;
  /** @nullable */
  amountSpent?: number | null;
  /** @nullable */
  costPerResult?: number | null;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  primaryText?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  callToAction?: string | null;
  /** @nullable */
  destinationUrl?: string | null;
  /** @nullable */
  mediaUrl?: string | null;
  /** @nullable */
  identityPage?: string | null;
  /** @nullable */
  instagramHandle?: string | null;
}