# lib/api-zod/src/generated/types/adSet.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Meta Ads Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AdSetBudgetType } from './adSetBudgetType';
import type { AdSetDelivery } from './adSetDelivery';
import type { AdSetStatus } from './adSetStatus';

export interface AdSet {
  id: number;
  campaignId: number;
  campaignName: string;
  name: string;
  status: AdSetStatus;
  delivery: AdSetDelivery;
  budgetType?: AdSetBudgetType;
  /** @nullable */
  budget?: number | null;
  amountSpent: number;
  impressions: number;
  reach: number;
  clicks: number;
  results: number;
  /** @nullable */
  costPerResult: number | null;
  /** @nullable */
  cpm: number | null;
  /** @nullable */
  ctr: number | null;
  /** @nullable */
  frequency: number | null;
  /** @nullable */
  audience?: string | null;
  /** @nullable */
  placement?: string | null;
  /** @nullable */
  optimization?: string | null;
  /** @nullable */
  billingEvent?: string | null;
  /** @nullable */
  conversionLocation?: string | null;
  /** @nullable */
  facebookPage?: string | null;
  /** @nullable */
  performanceGoal?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  ageMin?: number | null;
  /** @nullable */
  ageMax?: number | null;
  /** @nullable */
  languages?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  detailedTargeting?: string | null;
  /** @nullable */
  placementMode?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  recommendation?: string | null;
  createdAt: string;
}