# lib/api-zod/src/generated/types/adSetUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Meta Ads Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface AdSetUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  budgetType?: string | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  amountSpent?: number | null;
  /** @nullable */
  costPerResult?: number | null;
  /** @nullable */
  audience?: string | null;
  /** @nullable */
  conversionLocation?: string | null;
  /** @nullable */
  facebookPage?: string | null;
  /** @nullable */
  performanceGoal?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  ageMin?: number | null;
  /** @nullable */
  ageMax?: number | null;
  /** @nullable */
  languages?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  detailedTargeting?: string | null;
  /** @nullable */
  placementMode?: string | null;
  /** @nullable */
  placement?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}